refactor: move MCP tools into Tools module
Reorganized the MCP server tools into a dedicated Tools module for better code organization:
- Moved TaskTool, SessionInfoTool, and ResetSessionTool into lib/claude_swarm/tools/
- Added Tools namespace to all tool classes
- Updated ClaudeMcpServer to reference tools with the new namespace
- Updated all tests to use the new tool module structure

This change improves code organization and makes the tool structure more modular and maintainable.
